refactor(#572): extract LanguageResolver from SsrPageHandler
Extract language detection, negotiation, and path prefix stripping into
a dedicated LanguageResolver service. SsrPageHandler now accepts
LanguageResolver as an optional constructor dependency (auto-created
from serviceResolver if not provided). HttpKernel updated to call
getLanguageResolver() accessor. Tests updated to exercise LanguageResolver
directly for language-specific behavior.
