+++ docs/specs/infrastructure.md
@@ -2,7 +2,7 @@
 
 <!-- Spec reviewed 2026-04-05 - SovereigntyProfile/Config added to foundation, FoundationServiceProvider registers SovereigntyConfig singleton; CommunityContext/CommunityMiddleware added for community-scoped query isolation; SsrResponse removed, all controllers return Symfony Response/JsonResponse; ControllerDispatcher now delegates to DomainRouterInterface chain; both callable and router dispatch paths wrapped in try-catch returning 500 JSON:API errors; MediaRouter file move wrapped in try-catch; ViteAssetManager gained assetTags() method with devServerUrl constructor param for dev mode support; ControllerDispatcher uses Inertia::getRenderer() instead of hardcoded new RootTemplateRenderer(); RootTemplateRenderer accepts optional ViteAssetManager and injects Vite asset tags in default template; InertiaServiceProvider auto-configures renderer with ViteAssetManager for zero-config Inertia SPA support; AppControllerRouter added to dispatch Class::method controllers from ServiceProvider::routes() — delegates to SsrPageHandler::dispatchAppController, wired after SsrRouter in HttpKernel router chain (#1119); AppControllerRouter handle() relies on dispatchAppController's typed array shape contract (no runtime defensive casts); MediaRouter mkdir warning suppressed via @-prefix double-check idiom so a non-directory ancestor produces a clean 500 from the move catch block instead of a PHP warning under --fail-on-warning -->
 <!-- Spec reviewed 2026-04-06 - GraphQlRouter: inline `new \Waaseyaa\GraphQL\GraphQlEndpoint(...)` replaced with a proper `use Waaseyaa\GraphQL\GraphQlEndpoint` import (#1091 cleanup, no behavior change). Also reverted an incorrect fix that had added `waaseyaa/ssr` as a hard `require` of foundation's composer.json — that violated the layer rule (layer 0 must not depend on layer 6) and tripped `LayerDependencyTest::foundationDoesNotDependOnHigherLayerPackages`. The pre-existing architectural debt it was trying to paper over (HttpKernel directly imports `Waaseyaa\SSR\RenderCache`, `SsrPageHandler`, `SsrServiceProvider`, `TwigErrorPageRenderer`; `SsrRouter` and `AppControllerRouter` live in foundation but require `SsrPageHandler`; `EventListenerRegistrar::registerRenderCacheListeners()` type-hints `RenderCache`) is tracked as a separate P1 refactor follow-up to #571. -->
-<!-- Spec reviewed 2026-04-07 - ControllerDispatcher and DiscoveryRouter: import ordering corrected to satisfy PHP-CS-Fixer alphabetical rule (no behavior change). ControllerDispatcher now uses `use Waaseyaa\Inertia\Inertia` import instead of inline FQCN for `Inertia::getRenderer()` call. -->
+<!-- Spec reviewed 2026-04-07 - ControllerDispatcher and DiscoveryRouter: import ordering corrected to satisfy PHP-CS-Fixer alphabetical rule (no behavior change). ControllerDispatcher now uses `use Waaseyaa\Inertia\Inertia` import instead of inline FQCN for `Inertia::getRenderer()` call. LanguageResolver extracted from SsrPageHandler (#572): language detection, negotiation, and path prefix stripping now live in a dedicated service; HttpKernel delegates to SsrPageHandler::getLanguageResolver()->stripLanguagePrefixForRouting(). -->
 
 Specification for the foundational infrastructure layer of Waaseyaa CMS: domain events, cache system, database abstraction, query builder, migration system, kernel bootstrapping (including environment resolution and debug mode), service provider discovery, and queue workers.
