+++ packages/foundation/tests/Unit/Kernel/HttpKernelTest.php
@@ -20,6 +20,7 @@ use Waaseyaa\Database\DBALDatabase;
 use Waaseyaa\Entity\EntityTypeManager;
 use Waaseyaa\Foundation\Http\CorsHandler;
 use Waaseyaa\Api\Http\DiscoveryApiHandler;
+use Waaseyaa\SSR\LanguageResolver;
 use Waaseyaa\SSR\SsrPageHandler;
 use Waaseyaa\Foundation\Kernel\AbstractKernel;
 use Waaseyaa\Foundation\Kernel\BuiltinRouteRegistrar;
@@ -406,7 +407,7 @@ final class HttpKernelTest extends TestCase
     #[Test]
     public function render_language_resolution_uses_url_prefix_and_strips_alias_lookup_path(): void
     {
-        $handler = $this->createSsrPageHandler([
+        $resolver = $this->createLanguageResolver([
             'i18n' => [
                 'languages' => [
                     ['id' => 'en', 'label' => 'English', 'is_default' => true],
@@ -416,7 +417,7 @@ final class HttpKernelTest extends TestCase
         ]);
 
         $request = Request::create('/fr/teachings/water');
-        $resolved = $handler->resolveRenderLanguageAndAliasPath('/fr/teachings/water', $request);
+        $resolved = $resolver->resolveRenderLanguageAndAliasPath('/fr/teachings/water', $request);
 
         $this->assertSame('fr', $resolved['langcode']);
         $this->assertSame('/teachings/water', $resolved['alias_path']);
@@ -425,7 +426,7 @@ final class HttpKernelTest extends TestCase
     #[Test]
     public function render_language_resolution_uses_accept_language_when_no_url_prefix(): void
     {
-        $handler = $this->createSsrPageHandler([
+        $resolver = $this->createLanguageResolver([
             'i18n' => [
                 'languages' => [
                     ['id' => 'en', 'label' => 'English', 'is_default' => true],
@@ -436,7 +437,7 @@ final class HttpKernelTest extends TestCase
 
         $request = Request::create('/teachings/water');
         $request->headers->set('Accept-Language', 'fr-CA,fr;q=0.9,en;q=0.8');
-        $resolved = $handler->resolveRenderLanguageAndAliasPath('/teachings/water', $request);
+        $resolved = $resolver->resolveRenderLanguageAndAliasPath('/teachings/water', $request);
 
         $this->assertSame('fr', $resolved['langcode']);
         $this->assertSame('/teachings/water', $resolved['alias_path']);
@@ -445,10 +446,10 @@ final class HttpKernelTest extends TestCase
     #[Test]
     public function render_language_resolution_defaults_to_english_when_not_configured(): void
     {
-        $handler = $this->createSsrPageHandler();
+        $resolver = $this->createLanguageResolver();
 
         $request = Request::create('/teachings/water');
-        $resolved = $handler->resolveRenderLanguageAndAliasPath('/teachings/water', $request);
+        $resolved = $resolver->resolveRenderLanguageAndAliasPath('/teachings/water', $request);
 
         $this->assertSame('en', $resolved['langcode']);
         $this->assertSame('/teachings/water', $resolved['alias_path']);
@@ -833,6 +834,31 @@ final class HttpKernelTest extends TestCase
         array $config = [],
     ): \Waaseyaa\Foundation\Http\Router\MediaRouter {
         return new \Waaseyaa\Foundation\Http\Router\MediaRouter($projectRoot, $config);
+    }
+
+    private function createLanguageResolver(array $config = []): LanguageResolver
+    {
+        $languageDefs = $config['i18n']['languages'] ?? [
+            ['id' => 'en', 'label' => 'English', 'is_default' => true],
+        ];
+        $languages = array_map(
+            static fn(array $def) => new Language(
+                id: $def['id'],
+                label: $def['label'],
+                isDefault: $def['is_default'] ?? false,
+            ),
+            $languageDefs,
+        );
+        $manager = new LanguageManager($languages);
+
+        $serviceResolver = static function (string $className) use ($manager): ?object {
+            if ($className === LanguageManagerInterface::class) {
+                return $manager;
+            }
+            return null;
+        };
+
+        return new LanguageResolver(serviceResolver: $serviceResolver);
     }
 
     private function createSsrPageHandler(array $config = []): SsrPageHandler
