+++ phpstan-baseline.neon
@@ -498,7 +498,7 @@ parameters:
 		-
 			message: "#^Strict comparison using \\=\\=\\= between non\\-empty\\-array\\<int, string\\> and array\\{\\} will always evaluate to false\\.$#"
 			count: 1
-			path: packages/ssr/src/SsrPageHandler.php
+			path: packages/ssr/src/LanguageResolver.php
 
 		-
 			message: "#^Offset 'permissions' on array\\{label\\: string, permissions\\: array\\<int, string\\>\\} on left side of \\?\\? always exists and is not nullable\\.$#"
